feat: allow to skip destroying terraform created resources

+++ destroy_k8s_cluster.sh
@@ -2,8 +2,10 @@
 
 echo "Destroying kubernetes cluster"
 
-cd terraform
-terraform destroy --auto-approve
-cd ..
+if [[ ! " $* " =~ " --skip-terraform " ]]; then
+    cd terraform
+    terraform destroy --auto-approve
+    cd ..
+fi
 
 doctl kubernetes cluster delete k8s-experiments-cluster
